Add helper method to retrieve list of admin api features
