+++ Application/EdFi.Ods.Admin.Api/Infrastructure/WebApplicationBuilderExtensions.cs
@@ -1,6 +1,5 @@
 using System.Reflection;
 using EdFi.Admin.DataAccess.Contexts;
-using EdFi.Ods.Admin.Api.Features;
 using EdFi.Ods.Admin.Api.Infrastructure.Security;
 using EdFi.Ods.AdminApp.Management.Database;
 using EdFi.Security.DataAccess.Contexts;
@@ -126,18 +125,7 @@ public static class WebApplicationBuilderExtensions
 
     public static void AddFeatureSpecificServices(this IServiceCollection services)
     {
-        var featureInterface = typeof(IFeature);
-        var featureImpls = Assembly.GetExecutingAssembly().GetTypes()
-            .Where(p => featureInterface.IsAssignableFrom(p) && p.IsClass);
-
-        var features = new List<IFeature>();
-
-        foreach (var featureImpl in featureImpls)
-        {
-            if (Activator.CreateInstance(featureImpl) is IFeature feature)
-                features.Add(feature);
-        }
-        foreach (var routeBuilder in features)
+        foreach (var routeBuilder in Helpers.FeaturesHelper.GetFeatures())
         {
             routeBuilder.DefineFeatureSpecificServices(services);
         }
